Another bug
Fixed an issue where pawn moving up two spaces from initial rank on top of opponents piece.  It wasn't able to capture but was preventing a checkmate that should have been recognized.  I fixed this issue before but when i retyped the method to work for black and white pawns i forgot the line.