MoveFactory and AvailableMoves (New Classes)
As I mentioned I would create a class to store the available moves for all pieces.  This way getPossible<piece>Moves() would not be called twice.
Took validateMove() out of GameRules and put it in MoveFactory (new class).  It doesn't do validation anymore, just creates moves.  The validation is done through checking with the AvailableMoves class.
Also isCheckMateOrStaleMate() is now implemented in the Game class.  It is responsible for updating the AvailableMoves class after each move.  The Game object contains an AvailableMoves data field.

Lots of changes but I strongly believe it is for the better.